feat: Implement ClientPage component with loading state and search parameter handling
